refactor: migration from OpenAPIV3 to JSONSchema
