+++ components/tool/jsonschema/readme.md
@@ -1,6 +1,6 @@
 ## 说明
 
-eino 提供了把 openapi3 文档中的 json schema 转为 eino tool schema 的方法。
+eino 提供了把 json schema 转为 eino tool schema 的方法。
 对于想直接把一个 http 接口作为 eino 中的一个 tool 使用的场景，提供了便捷的方法。
 
 关于 tool 更多的信息，可以参考: https://www.cloudwego.io/zh/docs/eino/core_modules/components/tools_node_guide/
