--- components/tool/openapi3/openapi.json
@@ -1,44 +1,0 @@
-{
-    "openapi": "3.0.0",
-    "info": {
-      "title": "Todo API",
-      "version": "1.0.0"
-    },
-    "paths": {
-      "/api/v1/todo": {
-        "post": {
-          "summary": "Create a todo item",
-          "requestBody": {
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/TodoRequest"
-                }
-              }
-            }
-          },
-          "responses": {
-            "200": {
-              "description": "Todo item created"
-            }
-          }
-        }
-      }
-    },
-    "components": {
-      "schemas": {
-        "TodoRequest": {
-          "type": "object",
-          "properties": {
-            "title": {
-              "type": "string"
-            },
-            "completed": {
-              "type": "boolean"
-            }
-          },
-          "required": ["title"]
-        }
-      }
-    }
-  }
